# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: Dart Backend API
  version: 0.5.0
  description: Server-authoritative API for Dart | for you. Identity sessions use Secure HttpOnly cookies and CSRF tokens.
servers:
  - url: /api/v1
...
paths:
...
  /auth/register:
    post:
      tags: [Customer Auth]
      summary: Register a customer and queue a six-digit Email OTP
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerRegistration" }
      responses:
        "202":
          description: Email verification required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationRequired" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /auth/verify-email:
    post:
      tags: [Customer Auth]
      summary: Consume an Email OTP and create the customer session
      operationId: verifyCustomerEmail
      responses:
        "200": { $ref: "#/components/responses/Authenticated" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /auth/resend-verification:
    post:
      tags: [Customer Auth]
      summary: Replace a registration challenge and queue a new Email OTP
      operationId: resendCustomerVerification
      responses:
        "202": { description: Replacement verification challenge }
        "422": { $ref: "#/components/responses/ValidationError" }
  /auth/login:
    post:
      tags: [Customer Auth]
      summary: Create a customer session
      operationId: loginCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Login" }
      responses:
        "200": { $ref: "#/components/responses/Authenticated" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /representatives/login:
    post:
      tags: [Representative Auth]
      summary: Create a session for an approved representative
      operationId: loginRepresentative
      responses:
        "200": { $ref: "#/components/responses/Authenticated" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /auth/change-temporary-password:
    post:
      tags: [Customer Auth, Representative Auth]
      summary: Replace a temporary password and rotate all sessions
      operationId: changeTemporaryPassword
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "200": { $ref: "#/components/responses/Authenticated" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/refresh:
    post:
      tags: [Sessions]
      summary: Rotate the current server session
      operationId: refreshSession
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "200": { description: Rotated CSRF token and expiry }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      tags: [Sessions]
      summary: Revoke the current session
      operationId: logout
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "204": { description: Session revoked }
  /auth/logout-all:
    post:
      tags: [Sessions]
      summary: Revoke every session for the account
      operationId: logoutAll
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "204": { description: All sessions revoked }
  /auth/sessions:
    get:
      tags: [Sessions]
      summary: List active owned sessions
      operationId: listOwnSessions
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Active session list }
  /auth/sessions/{sessionId}:
    delete:
      tags: [Sessions]
      summary: Revoke one owned session
      operationId: revokeOwnSession
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Session revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /me:
    get:
      tags: [Sessions]
      summary: Return the minimum signed-in profile
      operationId: getMe
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Signed-in profile }
    patch:
      tags: [Customer Auth]
      summary: Update customer profile; changing email requires a new OTP
      operationId: updateMe
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "200": { description: Updated profile and optional verification challenge }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/representatives/{id}/approve:
    post:
      tags: [Staff Auth]
      summary: Approve a representative; Owner permission and MFA required
      operationId: approveRepresentative
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Representative approved }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/representatives/{id}/reject:
    post:
      tags: [Staff Auth]
      summary: Reject a representative; Owner permission and MFA required
      operationId: rejectRepresentative
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Representative rejected }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/password-reset-requests/{id}/temporary-password:
    post:
      tags: [Staff Auth]
      summary: Set a temporary password, revoke sessions and audit the action
      operationId: setTemporaryPassword
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Temporary password assigned }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dart_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    CustomerRegistration:
      type: object
      required: [name, email, phone1, password]
      properties:
        name: { type: string, minLength: 3, maxLength: 120 }
        email: { type: string, format: email }
        phone1: { type: string }
        phone2: { type: string }
        birthday: { type: string, format: date }
        password: { type: string, minLength: 4, writeOnly: true }
    Login:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string }
        password: { type: string, writeOnly: true }
        totp: { type: string, pattern: "^[0-9]{6}$", writeOnly: true }
...
    VerificationRequired:
      type: object
      required: [status, challengeId, expiresAt]
      properties:
        status: { type: string, const: verification_required }
        challengeId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
    Profile:
      type: object
      required: [id, accountType, code, name, email, phones, status]
      properties:
        id: { type: string, format: uuid }
        accountType: { type: string, enum: [customer, staff, representative] }
        code: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        phones: { type: array, items: { type: string } }
        birthday: { type: [string, "null"], format: date }
        status: { type: string }
        emailVerified: { type: boolean }
        mustChangePassword: { type: boolean }
    Authenticated:
      type: object
      required: [user, csrfToken]
      properties:
        user: { $ref: "#/components/schemas/Profile" }
        csrfToken: { type: string, writeOnly: true }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
  responses:
    Authenticated:
      description: Authenticated session created or replaced
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Authenticated" }
    Unauthorized:
      description: Authentication failed or session expired
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Duplicate or stale state conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Request or business validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...